pyproject.toml
[tool.poetry]
name = "recruitment-task"
version = "0.1.0"
description = ""
authors = ["MarcGroc <marcingrochola1@gmail.com>"]
readme = "README.md"
packages = [{include = "recruitment_task"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
python-dotenv = "^0.21.1"
pre-commit = "^3.0.4"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
flake8 = "^6.0.0"
black = "^23.1.0"
isort = "^5.12.0"
drf-spectacular = "^0.25.1"
pillow = "^9.4.0"
django-resized = "^1.0.2"
django-imagekit = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
